Make redis_backend.py reconnect on database failure
Currently if the connection to the database is lost, for any reason... Huey doesn't exit or reconnect to the database. It simply hangs.  
This change, while perhaps not the worlds best way to do it allows Huey to attempt to reconnect to the database should anything happen  

This includes socket timeout, database restart, network failure. All cases where connection is lost, Huey will attempt to reconnect.   
This change still allows for different database configurations for each class, and still attempts to only have a single active connection per class.

Let me know if you have a problem with it and I can attempt to do it another way.