Fix handling of negative sleep times.
In Python 2, calling time.sleep() with a negative argument has
platform-dependent results.  On Linux, it will throw an IOError, which is the
only case the code handles correctly.  On other platforms, it results in
sleeping for an infinite time, or undefined behaviour.  In Python 3,
time.sleep() will throw a ValueError when called with a negative argument.  See
https://bugs.python.org/issue12459 for additional information.

The easiest way to solve this problem for all Python versions is to compute the
time to sleep beforehand and only call time.sleep() if the argument is
postitive.  (This isn't in any way slower or less accurate than the previous
solution, as the comment I removed in this patch suggests.  CPython local
variables are accessed by static index rather than by name, and are thus just as
fast as the execution frame stack.)