Refactor signal handling for child processes. Ensure TERM handled.
Previously, when using worker processes, a worker that was busy
executing a task would not terminate immediately even when SIGTERM was
used. This meant that, in-effect, all shutdowns had to be graceful. This
issue is now resolved. When SIGTERM is sent, any workers executing a
task will be stopped (as originally intended). Graceful shutdown
continues to function as it always has.

Fixes #452